bugfix, pytest hidden state, marvelous
